Adding pre-build "minimal" and "basic" versions
In addition to the standard fabric.js build. This adds a "minimal" and "basic" version that can be required individually for cases where a more minimal version of Fabric is desired.

To use simply require( 'fabric-browseronly/minimal' ) to use the "minimal" version. Replace "minimal" with "basic" to use the basic version.